[bugfix] - shortest path was wrong
Shortest path was backtracing the wrong parents.
BFS was indeed working correctly as is stopped
as soon as shortest path came up.

The problem was that the parents where overwritten
every time a new node was added to the queue.

The solution is not to update a parent at later
insertions to the queue; if there a shortest path
to get to the node, then the new one is longer and
must not be recorded/remembered.

Change-Id: Id37660c66934e9c63f03f6da9fad2c6be3666beb

Nodes will only be added to queue if they are not
already present ( should avioid some unnecesary work )

Change-Id: Id99a2a10a82077acbc1e40e79a113725a01f33fd